fix: renaming the sensor files to lower case
