+++ measure.py
@@ -1,14 +1,15 @@
 import time
 import pslab
+import serial
 
 from sensors.ccs811_co2 import measure_co2
-from sensors.AO03_oxygen import measure_oxygen
-from sensors.GL5528_light import measure_light_intensity
-from sensors.HDC1080_temp import measure_temperature
+from sensors.ao03_oxygen import measure_oxygen
+from sensors.gl5528_light import measure_light_intensity
+from sensors.hdc1080_temp import measure_temperature
 
 MEASURING_INTERVAL = 1 # in seconds
 
-experiment_options = { # [function_name, unit]
+experiment_options = { # item_name : [function_name, unit]
     "co2"       : [measure_co2, "ppm"],
     "oxygen"    : [measure_oxygen, "%"],
     "light"     : [measure_light_intensity, "lux"],
@@ -16,7 +17,7 @@ experiment_options = { # [function_name, unit]
     }
     
 def connect_to_pslab():
-    """ Establishes the connection to the pslab. """
+    """ Establishes the connection to the PSLab. """
     try:
         psl = pslab.ScienceLab()
     except serial.SerialException: # device not found
@@ -33,7 +34,7 @@ def get_data_pslab(connection, experiment_type):
     while True:
         psl = connect_to_pslab()
         if psl is None:
-            return None
+            return None # failed
         while True:
             measurement = experiment_options[experiment_type][0](psl)
             if measurement != 0:
@@ -41,4 +42,4 @@ def get_data_pslab(connection, experiment_type):
                 time.sleep(MEASURING_INTERVAL)
             else:
                 time.sleep(10)
-                break # try to reconnect
+                break # try to reconnect - this works in the majority of times
