+++ shutdown.py
@@ -3,9 +3,8 @@ import os
 import time
 
 """
-Using a simple button to switch on/off.
-Connect one button "terminal" (both sides) with GPIO pin 27, the other one with GND next to it.
-Button has to be longpressed.
+Using a simple button to start the shutdown process of the Raspberry Pi.
+The goal is to give the user a chance to avoid hard resetting the Raspi every time.
 """
 
 PIN = 27
@@ -14,7 +13,7 @@ gpio.setup(PIN, gpio.IN)
 
 while True:
     if gpio.input(PIN) == gpio.LOW:
-        time.sleep(0.75)
+        time.sleep(0.5) # the button needs to be long-pressed to confirm the shutdown
         if gpio.input(PIN) is gpio.LOW:
             os.system("shutdown now-h")
             time.sleep(1)
