+++ store_data.py
@@ -9,7 +9,10 @@ FILEPATH = "/home/foss/data/"
 FLUSHING_THRESHOLD = 10
 
 def create_filename():
-	""" Creating a unique name for the new measurement file in the pattern of <file_path><date>(<seq_nr>).csv. """
+	"""
+	Creating a unique name for the new measurement file
+	in the pattern of <file_path><date>(<seq_nr>).csv.
+	"""
 	name = FILEPATH + time.strftime(DATESTAMP_FORMAT)
 	nr = 1
 	while os.path.isfile(name + ".csv"): # file with this name already exists
@@ -22,8 +25,8 @@ def write_to_csv(connection, experiment_type):
 	Writing the measurements line by line into a csv.
 	Flushing this output after a certain amount of lines (flushing_threshold),
 	so the user can follow the real time data on his device.
-	This also has the advantage, that a maximum of flushing_threshold-1 lines
-	of measurement are lost in the event of a sudden power outage.
+	This also has the advantage, that a maximum of |flushing_threshold-1| lines
+	of measurement are lost, in the event of a sudden power outage.
 	Each line has the format: <timestamp>, <measured_value>, <unit>, <item_name>.
 	"""
 	with open(create_filename(), "a") as f:
@@ -39,6 +42,6 @@ def write_to_csv(connection, experiment_type):
 				experiment_type
 				))
 			line_counter = line_counter + 1
-			if line_counter >= FLUSHING_THRESHOLD:
+			if line_counter >= FLUSHING_THRESHOLD: # flushing (aka updating) the csv now
 				f.flush()
 				line_counter = 0
